requirements.txt
# Core Backend Dependencies
fastapi==0.115.11
uvicorn==0.34.0          # ASGI server to run FastAPI

# OpenAI API Client
openai==0.28.0

# Web Framework for UI
streamlit==1.42.2

# HTTP Requests
requests==2.32.3

# JSON Handling & Data Processing
pydantic==2.10.6
python-dotenv==1.0.1
